Provides tail and handy ssh aliases
*tail* Powershell alias: this is missing from the standard
Powershell command-set. In a nutshell, this alias wraps the
command `Get-Content fileName -Wait`.

*connect server [profile]* this is the port of an alias I have
configured under Cmder for Bash. It is extremely handly to have
when one has access to a lot of Linux servers with SSHD enabled.
Inspired by the manner in which GoToFolder works, this alias is
flexible in the sense that users can define profiles allowing
them to assume different identities depending on the server
they need to access to. This is typically my use-case at work. The
alias comes with a "Usage" section and the example file is heavily
documented. Users are strongly encouraged to use Private / Public
key exchange for their SSHD authentication. It is a piece of tech
worth knowing about.

   --- Lionel Saliou, Ph.D